+++ worker/websocket.go
@@ -20,7 +20,7 @@ func randomID() int {
 	return 1e15 + int(rand.Float32()*2e15)
 }
 
-func connect(ctx context.Context, roomID int, out chan *message) error {
+func connect(ctx context.Context, roomID int64, out chan *message) error {
 	u := url.URL{
 		Scheme: "wss",
 		Host:   "broadcastlv.chat.bilibili.com:2245",
@@ -64,7 +64,7 @@ func connect(ctx context.Context, roomID int, out chan *message) error {
 
 // Connect is a blocking function that reads the message from broadcast with roomID and
 // then push it to the out channel
-func Connect(ctx context.Context, roomID int, out chan *message) {
+func Connect(ctx context.Context, roomID int64, out chan *message) {
 	err := connect(ctx, roomID, out)
 	if err != nil {
 		fmt.Printf("worker/websocket: [%d] %v\n", roomID, err)
